requirements.txt
feedparser
requests
python-dotenv
pydantic
openai
beautifulsoup4
html5lib
tqdm
langdetect

anthropic

pyyaml
